+++ src/components/content/conditions.tsx
@@ -23,7 +23,6 @@ import {
   createOrEditCondition,
   getAddConditionWithDefaults,
 } from "./forms/conditions";
-import { editPatientAndScheduleHistory } from "./forms/patients";
 import { PatientHistoryRequestDrawer } from "./patient-history-request-drawer";
 import { PatientHistoryMessage } from "./patient-history/patient-history-message";
 import {
@@ -79,7 +78,7 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
   const { getRequestContext } = useCTW();
   const [sort, setSort] = useState<TableSort>();
 
-  const [clinicalHistoryExists, setClinicalHistoryExists] = useState(false);
+  const [clinicalHistoryExists, setClinicalHistoryExists] = useState<boolean>();
 
   const patientRecordsMessage = patientRecordsResponse.isError
     ? ERROR_MSG
@@ -98,8 +97,6 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
     }
     return undefined;
   };
-
-  const handleToggleChange = () => setIncludeInactive(!includeInactive);
 
   const handleEditCondition = (condition: ConditionModel) => {
     if (patientResponse.data) {
@@ -173,7 +170,10 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
     </button>
   );
 
-  const handleClinicalHistory = async (patientID: string) => {
+  const shouldShowClinicalHistoryArea =
+    clinicalHistoryExists || otherProviderRecordsResponse.data?.length;
+
+  const checkClinicalHistory = async (patientID: string) => {
     const requestContext = await getRequestContext();
 
     const patientHistoryFetched = await hasFetchedPatientHistory(
@@ -181,12 +181,7 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
       patientID
     );
 
-    if (patientHistoryFetched) {
-      setClinicalHistoryExists(true);
-      setRequestDrawerIsOpen(false);
-    } else {
-      setClinicalHistoryExists(false);
-    }
+    setClinicalHistoryExists(patientHistoryFetched);
   };
 
   useEffect(() => {
@@ -218,15 +213,17 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
       }
     }
     void load();
-    if (patientResponse.data?.id) {
-      void handleClinicalHistory(patientResponse.data.id);
+    if (patientResponse.data?.id && clinicalHistoryExists === undefined) {
+      void checkClinicalHistory(patientResponse.data.id);
     }
+
     // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [
     includeInactive,
     patientResponse.data,
     patientRecordsResponse.data,
     otherProviderRecordsResponse.data,
+    clinicalHistoryExists,
     patientRecordsResponse.error,
   ]);
 
@@ -259,7 +256,7 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
           <div className="ctw-conditions-title-container">
             <div className="ctw-title">Patient Record</div>
             <ToggleControl
-              onFormChange={handleToggleChange}
+              onFormChange={() => setIncludeInactive(!includeInactive)}
               toggleProps={{ name: "conditions", text: "Include Inactive" }}
             />
           </div>
@@ -306,7 +303,7 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
         <div className="ctw-space-y-3">
           <div className="ctw-conditions-title-container">
             <div className="ctw-title">Other Provider Records</div>
-            {clinicalHistoryExists && (
+            {shouldShowClinicalHistoryArea && (
               <button
                 type="button"
                 className="ctw-btn-clear ctw-link"
@@ -316,7 +313,7 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
               </button>
             )}
           </div>
-          {clinicalHistoryExists ? (
+          {shouldShowClinicalHistoryArea ? (
             <ConditionsTableBase
               className="ctw-conditions-not-reviewed"
               stacked={breakpoints.sm}
@@ -388,7 +385,7 @@ export function Conditions({ className, readOnly = false }: ConditionsProps) {
           patient={patientResponse.data}
           isOpen={requestRecordsDrawerIsOpen}
           onClose={() => setRequestDrawerIsOpen(false)}
-          action={curry(editPatientAndScheduleHistory)(patientResponse.data)}
+          setClinicalHistoryExists={setClinicalHistoryExists}
         />
       )}